fix: 메인페이지 라우트 지정
